+++ tests/App/KernelTestCase.php
@@ -9,6 +9,7 @@ class KernelTestCase extends \Symfony\Bundle\FrameworkBundle\Test\KernelTestCase
 	protected static function bootKernel(array $options = []): KernelInterface {
 		static::ensureKernelShutdown();
 
+
         $kernel = new Kernel("test",$options['configDir'] ?? null, $options['preBoot'] ?? null);
 		$kernel->boot();
 		static::$kernel = $kernel;
